# crates/server/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Warg Registry API
  description: |
    [warg](https://warg.io/) is an open source protocol for WebAssembly component registries.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

# The Warg APIs currently do not use an authentication scheme
# Publishing requires that entries are signed with an acceptable key
...
servers:
  - url: http://localhost:8090/v1
    description: Local development server
paths:
...
  /package/{logId}/record:
    post:
      summary: Publish package record
      operationId: publishPackageRecord
      security: []
      tags:
        - package
      description: |
        Attempts to publish a new record to a package log.

        Publishing package records is an asynchronous operation.

        The record must be signed by a key that is authorized to modify the package log.
      parameters:
        - name: logId
          in: path
          description: The package log identifier.
          required: true
          schema:
            "$ref": "#/components/schemas/AnyHash"
        - name: Warg-Registry
          in: header
          description: URI Authority of another registry, if present and supported the registry will respond on behalf of another registry.
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishPackageRecordRequest"
      responses:
        "202":
          description: The package record was accepted.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PackageRecord"
        "403":
          description: |
            The key used to sign the record was not authorized to publish a record to the log.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: |
            The server does not support publishing package records with explicitly
            specified content source locations.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: An error occurred when processing the request.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /package/{logId}/record/{recordId}:
    get:
      summary: Get package record status
      operationId: getPackageRecord
      security: []
      tags:
        - package
      description: |
        Gets package record status from the registry.

        A package record is in one of the following states:
          * `sourcing`: The package record needs content sources.
          * `processing`: The package record is being processed.
          * `rejected`: The package record was rejected.
          * `published`: The package record was published to the log.
      parameters:
        - name: logId
          in: path
          description: The package log identifier.
          required: true
          schema:
            "$ref": "#/components/schemas/AnyHash"
        - name: recordId
          in: path
          description: The record identifier.
          required: true
          schema:
            "$ref": "#/components/schemas/AnyHash"
        - name: Warg-Registry
          in: header
          description: URI Authority of another registry, if present and supported the registry will respond on behalf of another registry.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The package record.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PackageRecord"
        "404":
          description: A requested entity was not found.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - status
                  - type
                  - id
                properties:
                  status:
                    type: integer
                    description: The HTTP status code for the error.
                    example: 404
                  type:
                    type: string
                    description: The type of entity that was not found.
                    enum: [log, record]
                    example: log
                  id:
                    "$ref": "#/components/schemas/AnyHash"
                    description: |
                      The identifier of the entity that was not found.
        default:
          description: An error occurred when processing the request.
          headers:
            Warg-Registry:
              schema:
                type: string
              description: URI Authority of another registry, if present and supported the registry responded on behalf of another registry.
              required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      description: A generic error response.
      additionalProperties: false
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: The HTTP status code for the error.
          example: 406
        message:
          type: string
          description: The error message.
          example: the server cannot produce an acceptable response
    AnyHash:
      type: string
      description: Represents a supported hash.
      example: sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c
      pattern: ^[a-z0-9-]+:[a-f0-9]+$
...
    PublishPackageRecordRequest:
      type: object
      description: A request to publish a record to a package log.
      additionalProperties: false
      required:
        - packageId
        - record
      properties:
        packageId:
          type: string
          description: The name of the package log being published to.
          maxLength: 128
          example: wasi:http
        record:
          "$ref": "#/components/schemas/EnvelopeBody"
          description: The package record being published to the log.
        contentSources:
          "$ref": "#/components/schemas/ContentSourceMap"
          description: |
            The map of all content sources for the record.

            A registry may not support specifying content sources for a record.

            If a registry does not support content sources, a 501 will be returned
            and content will need to be directly uploaded to the registry.
    PackageRecord:
      description: A package log record.
      allOf:
        - type: object
          required:
            - recordId
          properties:
            recordId:
              "$ref": "#/components/schemas/AnyHash"
              description: The record identifier.
        - oneOf:
            - "$ref": "#/components/schemas/SourcingRecord"
            - "$ref": "#/components/schemas/ProcessingRecord"
            - "$ref": "#/components/schemas/RejectedRecord"
            - "$ref": "#/components/schemas/PublishedRecord"
          discriminator:
            propertyName: state
            mapping:
              sourcing: "#/components/schemas/SourcingRecord"
              processing: "#/components/schemas/ProcessingRecord"
              rejected: "#/components/schemas/RejectedRecord"
              published: "#/components/schemas/PublishedRecord"
...
    SourcingRecord:
      type: object
      description: The package record is sourcing content.
      required:
        - state
        - missingContent
      properties:
        state:
          type: string
          description: The state of the package record.
          enum: [sourcing]
          example: sourcing
        missingContent:
          "$ref": "#/components/schemas/MissingContentMap"
          description: The missing content for the package record.
          minProperties: 1
    ProcessingRecord:
      type: object
      description: A record that is being processed.
      required:
        - state
      properties:
        state:
          type: string
          description: The state of the package record.
          enum: [processing]
          example: processing
    RejectedRecord:
      type: object
      description: A rejected package record.
      required:
        - state
        - reason
      properties:
        state:
          type: string
          description: The state of the package record.
          enum: [rejected]
          example: rejected
        reason:
          type: string
          description: The reason the package record was rejected.
          example: the first entry of the log is not `init`
    PublishedRecord:
      type: object
      description: A record that has been published to the log.
      required:
        - state
        - registryIndex
      properties:
        state:
          type: string
          description: The state of the package record.
          enum: [published]
          example: published
        registryIndex:
          type: integer
          description: The index of the record in the registry log.
...
    EnvelopeBody:
      description: A signed envelope body.
      allOf:
        - type: object
          required:
            - contentBytes
          properties:
            contentBytes:
              type: string
              description: |
                Base64-encoded bytes of the content.

                The content of an envelope body is typically a serialized protocol buffer
                representing an operator or package record.
              format: byte
              maxLength: 1048576
              example: "ZXhhbXBsZQ=="
        - $ref: "#/components/schemas/Signature"
...
    Signature:
      type: object
      description: Represents a signature of content.
      required:
        - keyId
        - signature
      properties:
        keyId:
          $ref: "#/components/schemas/AnyHash"
          description: The signing key identifier.
        signature:
          type: string
          description: The algorithm-prefixed bytes of the signature (base64 encoded).
          pattern: ^[a-z0-9-]+:(?:[A-Za-z0-9+\/]{4})*(?:[A-Za-z0-9+\/]{4}|[A-Za-z0-9+\/]{3}=|[A-Za-z0-9+\/]{2}={2})$
          example: "ecdsa-p256:MEUCIQCzWZBW6ux9LecP66Y+hjmLZTP/hZVz7puzlPTXcRT2wwIgQZO7nxP0nugtw18MwHZ26ROFWcJmgCtKOguK031Y1D0="
    MissingContentMap:
      type: object
      description: The map of content digest to missing content info.
      patternProperties:
        "^[a-z0-9-]+:[a-f0-9]+$":
          "$ref": "#/components/schemas/MissingContent"
      example:
        ? "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773"
        : upload:
            - type: http
              method: "POST"
              url: "https://example.com/7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773"
    MissingContent:
      description: Information about missing content.
      properties:
        upload:
          description: Upload endpoint(s) for the missing content.
          type: array 
          items:
            oneOf:
              - "$ref": "#/components/schemas/HttpUpload"
            discriminator:
              propertyName: type
              mapping:
                http: "#/components/schemas/HttpUpload"
    HttpUpload:
      type: object
      description: A HTTP upload endpoint.
      required:
        - type
        - method
        - url
      properties:
        type:
          type: string
          description: The type of upload endpoint.
          enum: [http]
          example: http
        method:
          type: string
          description: The HTTP method for the upload request. Uppercase is required.
          enum: [POST, PUT]
          example: POST
        url:
          type: string
          description: The URL of the upload endpoint, which may be relative to the API base URL.
          example: https://example.com/contents.wasm
          format: uri
        headers:
          type: object
          description: The HTTP headers for upload request. Header name is required to be all lowercase.
          properties:
            authorization:
              type: string
              description: Authorization header.
              example: "Bearer cn389ncoiwuencr"
            "content-type":
              type: string
              description: Content type header.
              example: "application/wasm"
...
    ContentSourceMap:
      type: object
      description: The map of content digest to sources.
      patternProperties:
        "^[a-z0-9-]+:[a-f0-9]+$":
          type: array
          items:
            "$ref": "#/components/schemas/ContentSource"
          description: The array of sources for the content digest.
          minItems: 1
          maxItems: 128
      example:
        ? "sha256:7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773"
        : - type: httpGet
            url: https://example.com/7d865e959b2466918c9863afca942d0fb89d7c9ac0c99bafc3749504ded9773
    ContentSource:
      description: A known content source.
      oneOf:
        - "$ref": "#/components/schemas/HttpGet"
      discriminator:
        propertyName: type
        mapping:
          httpGet: "#/components/schemas/HttpGet"
    HttpGet:
      type: object
      description: A known GET HTTP content source.
      required:
        - type
        - url
      properties:
        type:
          type: string
          description: The type of content source.
          enum: [httpGet]
          example: httpGet
        url:
          type: string
          description: The URL of the content.
          example: https://example.com/contents.wasm
          format: uri
        acceptRanges:
          type: boolean
          description: Flag indicating if the server accepts byte ranges with `Range` header.
        size:
          type: integer
          description: Content size in bytes.
          example: 1024
...